move SecondsNanos to new ts file, make internal
